This is the first shot at including code to create the pg cluster and create a database.
As noted in the README file, I am not familiar with rake, and so for this
first commit, the build uses a Makefile rather than a Rakefile.

The assumption is that the database will be set in the project root.  I
have used pgdata as the name of the database in the examples.  The port
used in the examples is 55432, and the database name is happiness.  These
are controlled via the PGDATA, PGPORT, and PGDATABASE environment variables.

See the README file for examples.